Datos sobre hechos armados ocurridos en Argentina entre el 26 de mayo de 1973 y el 24 de marzo de 1976.

Datos sobre hechos armados. Se trata de una serie de datos sobre hechos armados ocurridos en el país entre 1973 y 1976, relevados por un grupo de investigadorxs del CICSO.

Usage

hechos_armados

Format

se trata de una tabla con 8509 observaciones y 32 variables:

hecho

num número del hecho por año

fecha

POSIXct fecha del hecho

objeto

chr objeto del hecho

tipo

chr tipo de hecho

objeto1

chr contra quién, contra qué

fines

chr fines, objetivos

filiac

chr filiación partidaria u organizacional

filiac1

chr filiación de los objetos del hecho

atribuc

chr atribución de la responsabilidad del hecho

interv

chr intervención de otras fuerzas

enfrent

chr enfrentamineto

bajashum

chr tipo de bajas humanas de los participantes, objeto del hecho (el ‘atacado’)

totbajas

chr total de bajas humanas de los participantes objeto del hecho

bajas1

chr total de bajas humanas objeto del hecho que provoca el hecho, cuando el objeto no es humano

bajas2

chr tipo de bajas humanas entre los participantes ‘responsables’ del hecho (‘atacante’)

bajas3

chr total de bajas humanas de los participantes responsables del hecho

bajas4

chr tipo de bajas humanas de los participantes ‘intervinientes’ en el hecho (refuerzos)

bajas5

chr total de bajas humanas de los participantes intervinientes

bajas6

chr numero total de bajas humanas conocidas producidas en el hecho (excluidas las ‘ajenas’)

jerarq

chr jerarquía de las bajas de los participantes objeto del hecho

jerarq1

chr jerarquía de las bajas producidas a raíz del hecho, contra objeto no humano (‘atacado’)

jerarq2

chr si el objeto no es humano: jerarquía del dueño, habitante habitual, etc., que caracteriza al objeto humano

jerarq3

chr jerarquía de las bajas humanas de los participantes responsables del hecho (‘atacante’)

jerarq4

chr jerarquía de las bajas humanas de los participantes (refuerzos) intervinientes en el hecho

jerarq5

chr total de bajas humanas ajenas a los participantes objeto, responsables y /o intervinientes

material

chr bajas materiales gue se hayan producido a raíz del hecho

pertenen

chr pertenencia de las bajas materiales

ubicacio

chr ubicación del hecho: tipo de zona

departam

chr ubicación del hecho: por departamento

provinci

chr ubicación del hecho: por provincia

ciudad

chr ubicación del hecho: por ciudad

zonas

chr tipo de zona urbana

Source

Marín, J. C. (1984) Los hechos armados un ejercicio posible. Centro de Investigaciones en Ciencias Sociales. Buenos Aires.